Conflicts:
	py_gnome/tests/unit_tests/test_model.py
	py_gnome/tests/unit_tests/test_random_mover.py
	py_gnome/tests/unit_tests/test_spill.py
	py_gnome/tests/unit_tests/test_spill_container.py
	py_gnome/tests/unit_tests/test_utilities/test_cache.py
	py_gnome/tests/unit_tests/test_wind_mover.py

(WIP) Also updated OilProps (in spill.py) init so it creates the OilLibrary.db
if it does not exist. This is just a placeholder for now, we may want to
do it differently in the long term.

+++ lib_gnome/RiseVelocity_c.h
@@ -16,25 +16,33 @@
 #include "ExportSymbols.h"
 
 class DLL_API RiseVelocity_c : virtual public Mover_c {
-	
+
 public:
 	double water_density;
 	double water_viscosity;
-	
+
 #ifndef pyGNOME
 	RiseVelocity_c (TMap *owner, char *name);
 #endif
 	RiseVelocity_c();
-	virtual OSErr 		PrepareForModelRun(); 
-	virtual OSErr 		PrepareForModelStep(const Seconds&, const Seconds&, bool, int numLESets, int* LESetsSizesList); 
-	virtual void 		ModelStepIsDone();
-	virtual WorldPoint3D       GetMove(const Seconds& model_time, Seconds timeStep,long setIndex,long leIndex,LERec *theLE,LETYPE leType);
-	
-	
-	OSErr				get_move(int n, unsigned long model_time, unsigned long step_len, WorldPoint3D* ref, WorldPoint3D* delta, double* rise_velocity, double* density, long* droplet_size, short* LE_status, LEType spillType, long spill_ID);
+
+	virtual OSErr PrepareForModelRun();
+	virtual OSErr PrepareForModelStep(const Seconds&, const Seconds&, bool, int numLESets, int* LESetsSizesList);
+
+	virtual void ModelStepIsDone();
+
+	virtual WorldPoint3D GetMove(const Seconds& model_time, Seconds timeStep,
+								 long setIndex, long leIndex, LERec *theLE, LETYPE leType);
+
+	OSErr get_move(int n, unsigned long model_time, unsigned long step_len,
+				   WorldPoint3D *ref, WorldPoint3D *delta,
+				   double *rise_velocity, double *density, double *droplet_size,
+				   short *LE_status, LEType spillType, long spill_ID);
 
 protected:
-	void				Init();
+
+	void Init();
+
 };
 
 #endif
